Move BaseElement into mathics.core.element
