+++ mathics/core/pattern.py
@@ -3,8 +3,9 @@
 # -*- coding: utf-8 -*-
 
 
+from mathics.core.element import ensure_context
 from mathics.core.expression import Expression
-from mathics.core.symbols import system_symbols, ensure_context, Atom, Symbol
+from mathics.core.symbols import Atom, Symbol, system_symbols
 from mathics.core.util import subsets, subranges, permutations
 from itertools import chain
